requirements.txt
streamlit==1.29.0
langchain-groq==0.1.0
sentence-transformers==2.6.1
torch==2.1.2
PyMuPDF==1.23.3
requests==2.31.0
pandas==2.2.0
python-dotenv==1.0.0
networkx==3.2.1
pyvis==0.3.2
openai==1.12.0
arxiv==2.1.0
tqdm==4.66.2
numpy==1.26.3
typing-extensions==4.9.0
lightrag-hku==0.0.7
lightrag @ git+https://github.com/your_username/lightrag.git
torch==2.1.2+cu118 -f https://download.pytorch.org/whl/torch_stable.html